requirements.txt
Django
djangorestframework
django-cors-headers
pika
celery
django-celery-beat
requests
psycopg2
Redis